+++ loan-management-service/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [0.0.2] - 2026-09-03
 ## Added
-    - added create laons and schedule payments functional  
+    - added create loans and schedule payments functional  
     - added process payments REGULAR and FULL
 
 [0.0.1] - 2026-09-01